refactor(app-agent): extract intent action display types
